ai-service/requirements.txt
# Core AI Framework
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.20

# LLM Inference - GPT-OSS via llama.cpp or Ollama
llama-cpp-python>=0.2.0
langchain-ollama>=0.1.0

# API & Streaming
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=13.0,<14
python-multipart>=0.0.6

# gRPC for kernel communication
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Data & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0

